+++ resources/views/components/boolean.blade.php
@@ -1,10 +1,10 @@
 <input 
-    name="{{ $column->getName() }}" 
+    name="{{ $column['name'] }}" 
     value="0" 
     type="hidden" />
 
 <input 
-    name="{{ $column->getName() }}" 
+    name="{{ $column['name'] }}" 
     value="1" 
     type="checkbox"
     {{ $value ? 'checked' : '' }} />