+++ resources/views/components/date.blade.php
@@ -1,4 +1,4 @@
 <input 
-    name="{{ $column->getName() }}" 
+    name="{{ $column['name'] }}" 
     value="{{ $value ?? date('Y-m-d', \strtotime($value)) }}" 
     type="date" />