+++ resources/views/components/datetime-local.blade.php
@@ -1,4 +1,4 @@
 <input 
-    name="{{ $column->getName() }}" 
+    name="{{ $column['name'] }}" 
     value="{{ date('Y-m-d\TH:i', \strtotime($value)) }}" 
     type="datetime-local" />