+++ resources/views/components/decimal.blade.php
@@ -1,5 +1,5 @@
 <input 
-    name="{{ $column->getName() }}" 
+    name="{{ $column['name'] }}" 
     value="{{ $value }}" 
     type="number"
     step="any" />